+++ Assets/ParticleNetwork/Scripts/Modules/Core/Utils/ChainUtils.cs
@@ -5,7 +5,7 @@ namespace Network.Particle.Scripts.Core.Utils
 {
     public class ChainUtils
     {
-        public static ChainInfo CreateChain(string chainName, int chainId)
+        public static ChainInfo CreateChain(string chainName, long chainId)
         {
             var chainInfo = (ChainInfo)Activator.CreateInstance(
                 Type.GetType($"Network.Particle.Scripts.Model.{chainName}Chain"),
